+++ R/era-package.R
@@ -12,6 +12,11 @@ NULL
 #' @name era-vctrs
 NULL
 
+#' @importFrom zeallot %<-%
+#' @export
+#' @keywords internal
+NULL
+
 # The following block is used by usethis to automatically manage
 # roxygen namespace tags. Modify with care!
 ## usethis namespace: start